adding all exception handlers needed and used DTO's validation tools

+++ src/main/java/com/global/hr/admin/service/AdminService.java
@@ -1,5 +1,10 @@
 package com.global.hr.admin.service;
 
+import com.global.hr.admin.DTO.AdminRequest;
+import com.global.hr.admin.DTO.AdminResponse;
+import com.global.hr.admin.Exception.AdminNotFoundException;
+import com.global.hr.admin.Exception.DoctorNotFoundException;
+import com.global.hr.admin.Exception.DuplicateEntryException;
 import com.global.hr.admin.entity.Admin;
 import com.global.hr.admin.repository.AdminRepository;
 import com.global.hr.doctor.entity.Doctor;
@@ -9,8 +14,12 @@ import com.global.hr.doctor.service.DoctorService;
 import com.global.hr.doctor.service.ExcuseService;
 import com.global.hr.patient.DataTransferObjects.PatientRequest;
 import com.global.hr.patient.DataTransferObjects.PatientResponse;
+import com.global.hr.patient.Exception.ResourceNotFoundException;
+import com.global.hr.patient.Models.Patient;
 import com.global.hr.patient.Services.AppointmentServices;
 import com.global.hr.patient.Services.PatientServices;
+
+import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
@@ -21,17 +30,20 @@ public class AdminService {
     private final AdminRepository adminRepo;
     private final DoctorService doctorService;
     private final PatientServices patientService;
-    public final AvailableTimeService avalable_time_service;
-    public final ExcuseService excuseService;
-    public final AppointmentServices appointmentService;
+    private final AvailableTimeService avalable_time_service;
+    private final ExcuseService excuseService;
+    private final AppointmentServices appointmentService;
+    private final ModelMapper modelMapper;
 
-    public AdminService(AdminRepository adminRepo, DoctorService doctorService, PatientServices patientService, AvailableTimeService avalable_time_service, ExcuseService excuseService, AppointmentServices appointmentService) {
+
+    public AdminService(AdminRepository adminRepo, DoctorService doctorService, PatientServices patientService, AvailableTimeService avalable_time_service, ExcuseService excuseService, AppointmentServices appointmentService, ModelMapper modelMapper) {
         this.adminRepo = adminRepo;
         this.doctorService = doctorService;
         this.patientService = patientService;
         this.avalable_time_service = avalable_time_service;
         this.excuseService = excuseService;
         this.appointmentService = appointmentService;
+        this.modelMapper = modelMapper;
     }   
 
     // ===== Admin CRUD =====
@@ -41,20 +53,33 @@ public class AdminService {
         return admin != null && admin.getPassword().equals(password);
     }
     
-    public Admin createAdmin(Admin admin) {
-        return adminRepo.save(admin);
+    public AdminResponse createAdmin(AdminRequest admin_req) {
+        if (adminRepo.existsByUsername(admin_req.getUsername())){
+            throw new DuplicateEntryException("Username already exists");
+        }
+        Admin admin = modelMapper.map(admin_req, Admin.class);
+        Admin new_admin =  adminRepo.save(admin);
+        return modelMapper.map(new_admin, AdminResponse.class);
     }
 
-    public List<Admin> getAllAdmins() {
-        return adminRepo.findAll();
+    public List<AdminResponse> getAllAdmins() {
+        List<Admin> all_admins = adminRepo.findAll();
+        return modelMapper.map(all_admins, new org.modelmapper.TypeToken<List<AdminResponse>>(){}.getType());
     }
 
     public void deleteAdmin(int id) {
-        adminRepo.deleteById(id);
+       if(!adminRepo.existsById(id)){
+            throw new AdminNotFoundException("Couldn't delete the admin as no admin with id: "+ id + " is found in the database.");
+       }
+       else{
+           adminRepo.deleteById(id);
+
+       }
     }
 
     // ===== Doctor Management =====
     public Doctor addDoctor(Doctor doctor) {
+       
         return doctorService.saveDoctor(doctor);
     }
 
@@ -63,6 +88,11 @@ public class AdminService {
     }
 
     public void deleteDoctor(int id) {
+
+        Optional<Doctor> doctor = doctorService.getDoctorById(id);
+        if (doctor.isEmpty()){
+            throw new DoctorNotFoundException("Couldn't delete the doctor as no doctor with id: "+ id + " is found in the database.");
+        }
         avalable_time_service.deleteAllRecordsByDoctorId(id);
         excuseService.deleteAllRecordsByDoctorId(id);
         appointmentService.deleteAllRecordsByDoctorId(id);
@@ -72,6 +102,12 @@ public class AdminService {
 
     // ===== Patient Management =====
     public PatientResponse addPatient(PatientRequest request) {
+        Patient patient = modelMapper.map(request, Patient.class);
+        Optional<Patient> savedPatient = patientService.getPatientRepository().findByEmailAndName(patient.getEmail(), patient.getName());
+        if (savedPatient.isPresent()){
+            throw new DuplicateEntryException("Patient with name: "+ patient.getName() + " and email: "+ patient.getEmail() + " is already present.");
+        }
+				
         return patientService.signUp(request);
     }
